refactor:  separate routes and segregate it on its own file. the reason is the index route become bloated and hard to  find specific route
