+++ src/router/index.js
@@ -1,112 +1,9 @@
-import GuestLayout from '@/views/Layouts/GuestLayout.vue';
-import OrganizationLayout from '@/views/Layouts/OrganizationLayout.vue';
 import { createRouter, createWebHistory } from 'vue-router'
- 
-function auth(to, from, next) {
-  if (!localStorage.getItem("access_token")) {
-    return next({ name: "login" });
-  }
- 
-  next();
-}
-
-function guest(to, from, next) {
-  if (localStorage.getItem("access_token")) {
-    return next({ name: "organization.dashboard" });
-  }
- 
-  next();
-}
+import guestRoutes from '@/router/guest';
+import authRoutes from '@/router/auth';
 const router = createRouter({
   history: createWebHistory(import.meta.env.BASE_URL),
-  routes: [
-    {
-      path: '/',
-      name: 'guest',
-      component: GuestLayout,
-      beforeEnter: guest,
-      children: [
-        {
-          path: '/',
-          name: 'welcome',
-          component: () => import('@/views/HomeView.vue'),
-        },
-        {
-          path: '/register',
-          name: 'register',
-          component: () => import('@/views/Auth/RgisterView.vue'),
-        },
-        {
-          path: '/login',
-          name: 'login',
-          component: () => import('@/views/Auth/LoginView.vue'),
-        }
-      ]
-    },
-    {
-      path: '/',
-      name: 'organization',
-      component: OrganizationLayout,
-      beforeEnter: auth,
-      children: [
-        {
-          path: '/organization/dashboard',
-          name: 'organization.dashboard',
-          component: () => import('@/views/Dashboards/OrganizationDashboard.vue'),
-        },
-        {
-          path: '/profile',
-          name: 'profile.index',
-          component: () => import('@/views/Auth/ProfileView.vue'),
-        },
-        {
-          path: '/projects/:projectId/tasks/create',
-          name: 'tasks.create',
-          component: () => import('@/views/Tasks/CreateView.vue'),
-        },
-        {
-          path: 'projects/:projectId/tasks/:taskId',
-          name: 'tasks.show',
-          component: () => import('@/views/Tasks/ShowView.vue'),
-        },
-        {
-          path: '/members',
-          name: 'members.index',
-          component: () => import('@/views/Members/IndexView.vue'),
-        },
-        {
-          path: '/members/create',
-          name: 'members.create',
-          component: () => import('@/views/Members/CreateView.vue'),
-        },
-        {
-          path: '/members/:id',
-          name: 'members.show',
-          component: () => import('@/views/Members/ShowView.vue'),
-        },
-        {
-          path: '/projects',
-          name: 'projects.index',
-          component: () => import('@/views/Projects/IndexView.vue'),
-        },
-        {
-          path: '/projects/create',
-          name: 'projects.create',
-          component: () => import('@/views/Projects/CreateView.vue'),
-        },
-        {
-          path: '/projects/:projectId',
-          name: 'projects.show',
-          component: () => import('@/views/Projects/ShowView.vue'),
-        },
-        {
-          path: '/projects/:projectId/edit',
-          name: 'projects.edit',
-          component: () => import('@/views/Projects/EditView.vue'),
-        },
-      ]
-    },
-  ],
+  routes: [guestRoutes, authRoutes],
 })
  
 export default router
