+++ src/router/team.js
@@ -1,0 +1,7 @@
+export default [
+    {
+        path: '/teams',
+        name: 'teams.index',
+        component: () => import('@/views/Teams/IndexView.vue'),
+    },
+  ];
